Corrected license information
By converting the original SID emulator into an arduino library I found the wording of the LGPL to better match what it was now, a library instead of a program. By doing this I inadvertedly changed the license conditions into a more permissive one.
I contacted the author and informed him about his. He gave me his blessing to keep the LGPL but actually  the code was GPL v2 based and any change is actually non-effective unless it is contemplated by the license.
Any derivative work must attach to the the GPL v2 or any later version. This commit is to be included in any fork.